More test helper sharing
